requirements.txt
functions-framework==3.*
google-cloud-storage
pillow
opencv-python
numpy
tensorflow==2.15.0
matplotlib
requests
firebase-admin